requirements.txt
mcp[cli]>=1.3.0
openai>=1.0.0
numpy>=1.24.0
jinja2>=3.1.0
weasyprint>=62.0
honcho-ai>=2.0.1
